move results_controller#create to api::results_controller#create
