+++ spec/spec_helper.rb
@@ -86,3 +86,5 @@ RSpec.configure do |config|
   Kernel.srand config.seed
 =end
 end
+
+require 'shared_api.rb'
